+++ source/_static/js/redirects.js
@@ -94,6 +94,8 @@ newUrls['4.14'] = [
   '/user-manual/api/requests-responses.html',
   '/user-manual/api/api-examples.html',
   '/user-manual/api/rbac/configuration.html',
+  '/user-manual/wazuh-server-cluster/adding-new-server-nodes/all-in-one-deployment.html',
+  '/user-manual/wazuh-server-cluster/adding-new-server-nodes/distributed-deployment.html',
 ]
 
 /* Redirections 4.13—4.14  */
@@ -294,6 +296,56 @@ redirections.push(
     '4.13': '/user-manual/api/rbac/how-it-works.html#priority-of-roles-and-policies',
     '4.14': '/user-manual/api/rbac/configuration.html#priority-of-roles-and-policies',
   },
+  {
+    'target': ['4.13=>4.14'],
+    '4.13': '/user-manual/wazuh-server-cluster/adding-new-server-nodes/certificates-creation.html',
+    '4.14': '/user-manual/wazuh-server-cluster/adding-new-server-nodes/index.html',
+  },
+  {
+    'target': ['4.13=>4.14'],
+    '4.13': '/user-manual/wazuh-server-cluster/adding-new-server-nodes/distributed-deployment.html',
+    '4.14': '/user-manual/wazuh-server-cluster/adding-new-server-nodes/index.html',
+  },    
+  {
+    'target': ['4.13=>4.14', '4.14=>4.13'],
+    '4.13': '/user-manual/wazuh-server-cluster/adding-new-server-nodes/certificates-creation.html#all-in-one-deployment',
+    '4.14': '/user-manual/wazuh-server-cluster/adding-new-server-nodes/all-in-one-deployment.html#certificate-creation',
+  },
+  {
+    'target': ['4.13=>4.14', '4.14=>4.13'],
+    '4.13': '/user-manual/wazuh-server-cluster/adding-new-server-nodes/configuration-to-connect-with-new-node.html#all-in-one-deployment',
+    '4.14': '/user-manual/wazuh-server-cluster/adding-new-server-nodes/all-in-one-deployment.html#configuring-existing-components-to-connect-with-the-new-node',
+  },
+  {
+    'target': ['4.13=>4.14', '4.14=>4.13'],
+    '4.13': '/user-manual/wazuh-server-cluster/adding-new-server-nodes/server-nodes-installation.html#all-in-one-deployment',
+    '4.14': '/user-manual/wazuh-server-cluster/adding-new-server-nodes/all-in-one-deployment.html#wazuh-server-worker-node-s-installation',
+  },
+  {
+    'target': ['4.13=>4.14', '4.14=>4.13'],
+    '4.13': '/user-manual/wazuh-server-cluster/adding-new-server-nodes/testing-the-cluster.html#all-in-one-deployment',
+    '4.14': '/user-manual/wazuh-server-cluster/adding-new-server-nodes/all-in-one-deployment.html#testing-the-cluster',
+  },
+  {
+    'target': ['4.13=>4.14', '4.14=>4.13'],
+    '4.13': '/user-manual/wazuh-server-cluster/adding-new-server-nodes/certificates-creation.html#distributed-deployment',
+    '4.14': '/user-manual/wazuh-server-cluster/adding-new-server-nodes/distributed-deployment.html#certificate-creation',
+  },
+  {
+    'target': ['4.13=>4.14', '4.14=>4.13'],
+    '4.13': '/user-manual/wazuh-server-cluster/adding-new-server-nodes/configuration-to-connect-with-new-node.html#distributed-deployment',
+    '4.14': '/user-manual/wazuh-server-cluster/adding-new-server-nodes/distributed-deployment.html#configuring-existing-components-to-connect-with-the-new-node',
+  },
+  {
+    'target': ['4.13=>4.14', '4.14=>4.13'],
+    '4.13': '/user-manual/wazuh-server-cluster/adding-new-server-nodes/server-nodes-installation.html#distributed-deployment',
+    '4.14': '/user-manual/wazuh-server-cluster/adding-new-server-nodes/distributed-deployment.html#wazuh-server-worker-node-s-installation',
+  },
+  {
+    'target': ['4.13=>4.14', '4.14=>4.13'],
+    '4.13': '/user-manual/wazuh-server-cluster/adding-new-server-nodes/testing-the-cluster.html#distributed-deployment',
+    '4.14': '/user-manual/wazuh-server-cluster/adding-new-server-nodes/distributed-deployment.html#testing-the-cluster',
+  },
 );
 
 /* Pages no longer available in 4.14 (this is not required if there is a redirection for this url) */
@@ -325,6 +377,10 @@ removedUrls['4.14'] = [
   '/user-manual/user-administration/single-sign-on/read-only/onelogin.html',
   '/user-manual/user-administration/single-sign-on/read-only/pingone.html',
   '/user-manual/agent/agent-management/index.html',
+  '/user-manual/wazuh-server-cluster/adding-new-server-nodes/certificates-creation.html',
+  '/user-manual/wazuh-server-cluster/adding-new-server-nodes/configuration-to-connect-with-new-node.html',
+  '/user-manual/wazuh-server-cluster/adding-new-server-nodes/server-nodes-installation.html',
+  '/user-manual/wazuh-server-cluster/adding-new-server-nodes/testing-the-cluster.html',
 ];
 
 /* *** RELEASE 4.13 ****/